+++ packages/mochi-ui/src/components/txns.ts
@@ -27,17 +27,8 @@ type Props = {
   global?: boolean;
 };
 
-function isVault(
-  tx: VaultTransferTx | TransferTx,
-  source: "from" | "to"
-): tx is VaultTransferTx {
-  return (
-    (source === "from" ? tx.from_profile_source : tx.other_profile_source) ===
-      "mochi-vault" &&
-    "metadata" in tx &&
-    "vault_request" in tx.metadata &&
-    "vault_id" in tx.metadata.vault_request
-  );
+function isVault(tx: VaultTransferTx | TransferTx): tx is VaultTransferTx {
+  return tx.action === "vault_transfer";
 }
 
 export async function formatTxn(
@@ -84,6 +75,7 @@ export async function formatTxn(
 
   if ("action" in tx) {
     switch (tx.action) {
+      case "vault_transfer":
       case "transfer": {
         switch (tx.type) {
           case "in": {
@@ -103,13 +95,13 @@ export async function formatTxn(
             if (global) {
               const [from, to] = await UI.resolve(
                 on,
-                isVault(tx, "from")
+                isVault(tx)
                   ? {
                       type: "vault",
                       id: tx.metadata.vault_request.vault_id.toString(),
                     }
                   : tx.from_profile_id,
-                isVault(tx, "to")
+                isVault(tx)
                   ? {
                       type: "vault",
                       id: tx.metadata.vault_request.vault_id.toString(),
@@ -122,7 +114,7 @@ export async function formatTxn(
             } else {
               const [from] = await UI.resolve(
                 on,
-                isVault(tx, "to")
+                isVault(tx)
                   ? {
                       type: "vault",
                       id: tx.metadata.vault_request.vault_id.toString(),
@@ -153,13 +145,13 @@ export async function formatTxn(
             if (global) {
               const [from, to] = await UI.resolve(
                 on,
-                isVault(tx, "from")
+                isVault(tx)
                   ? {
                       type: "vault",
                       id: tx.metadata.vault_request.vault_id.toString(),
                     }
                   : tx.from_profile_id,
-                isVault(tx, "to")
+                isVault(tx)
                   ? {
                       type: "vault",
                       id: tx.metadata.vault_request.vault_id.toString(),
@@ -172,7 +164,7 @@ export async function formatTxn(
             } else {
               const [to] = await UI.resolve(
                 on,
-                isVault(tx, "to")
+                isVault(tx)
                   ? {
                       type: "vault",
                       id: tx.metadata.vault_request.vault_id.toString(),